Rename package inspectors.metrics4bears to inspectors.properties
