+++ repairnator/repairnator-pipeline/src/main/java/fr/inria/spirals/repairnator/process/inspectors/properties/commits/Commit.java
@@ -1,4 +1,4 @@
-package fr.inria.spirals.repairnator.process.inspectors.metrics4bears.commits;
+package fr.inria.spirals.repairnator.process.inspectors.properties.commits;
 
 import java.util.Date;
