+++ repairnator/repairnator-pipeline/src/main/java/fr/inria/spirals/repairnator/process/inspectors/properties/reproductionBuggyBuild/ProcessDurations.java
@@ -1,4 +1,4 @@
-package fr.inria.spirals.repairnator.process.inspectors.metrics4bears.reproductionBuggyBuild;
+package fr.inria.spirals.repairnator.process.inspectors.properties.reproductionBuggyBuild;
 
 import fr.inria.spirals.repairnator.process.step.AbstractStep;
 import fr.inria.spirals.repairnator.process.step.BuildProject;
