+++ repairnator/repairnator-pipeline/src/main/java/fr/inria/spirals/repairnator/process/inspectors/properties/tests/FailingClass.java
@@ -1,19 +1,20 @@
-package fr.inria.spirals.repairnator.process.inspectors.metrics4bears.tests;
+package fr.inria.spirals.repairnator.process.inspectors.properties.tests;
 
-import java.util.HashSet;
-import java.util.Set;
+public class FailingClass {
 
-public class OverallMetrics {
-
+    private String testClass;
     private int numberRunning;
     private int numberPassing;
     private int numberFailing;
     private int numberErroring;
     private int numberSkipping;
-    private Set<Failure> failures;
 
-    public OverallMetrics() {
-        this.failures = new HashSet<>();
+    public FailingClass(String testClass) {
+        this.testClass = testClass;
+    }
+
+    public String getTestClass() {
+        return testClass;
     }
 
     public int getNumberRunning() {
@@ -55,22 +56,4 @@ public class OverallMetrics {
     public void setNumberSkipping(int numberSkipping) {
         this.numberSkipping = numberSkipping;
     }
-
-    public Set<Failure> getFailures() {
-        return failures;
-    }
-
-    public void addFailure(String failureName, boolean isError) {
-        for (Failure failure : this.failures) {
-            if (failure.getFailureName().equals(failureName) &&
-                    failure.getIsError() == isError) {
-                failure.setOccurrences(failure.getOccurrences() + 1);
-                return;
-            }
-        }
-        Failure failure = new Failure(failureName, isError);
-        failure.setOccurrences(1);
-        this.failures.add(failure);
-    }
-
 }
