+++ repairnator/repairnator-pipeline/src/main/java/fr/inria/spirals/repairnator/process/inspectors/properties/tests/FailureDetail.java
@@ -1,4 +1,4 @@
-package fr.inria.spirals.repairnator.process.inspectors.metrics4bears.tests;
+package fr.inria.spirals.repairnator.process.inspectors.properties.tests;
 
 public class FailureDetail {
