+++ repairnator/repairnator-pipeline/src/main/java/fr/inria/spirals/repairnator/process/step/checkoutrepository/CheckoutRepository.java
@@ -9,8 +9,8 @@ import fr.inria.spirals.repairnator.process.inspectors.JobStatus;
 import fr.inria.spirals.repairnator.process.inspectors.Metrics;
 import fr.inria.spirals.repairnator.process.inspectors.ProjectInspector;
 import fr.inria.spirals.repairnator.process.inspectors.StepStatus;
-import fr.inria.spirals.repairnator.process.inspectors.metrics4bears.commits.Commit;
-import fr.inria.spirals.repairnator.process.inspectors.metrics4bears.Metrics4Bears;
+import fr.inria.spirals.repairnator.process.inspectors.properties.commits.Commit;
+import fr.inria.spirals.repairnator.process.inspectors.properties.Metrics4Bears;
 import fr.inria.spirals.repairnator.process.step.AbstractStep;
 import fr.inria.spirals.repairnator.states.PipelineState;
 import org.eclipse.jgit.api.Git;
